Split models into their own file
