+++ Predicting.py
@@ -1,11 +1,8 @@
 import csv
 from sklearn.preprocessing import StandardScaler
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.cluster import KMeans
-from sklearn.neural_network import MLPClassifier
 from sklearn.metrics import accuracy_score
 from sklearn.decomposition import PCA
-
+import Models.models as models
 
 def get_games(file_path: str) -> list:
     games = []
@@ -16,13 +13,11 @@ def get_games(file_path: str) -> list:
                 games.append(row)
     return games
 
-
 def get_independent(games: list) -> list:
     x = []
     for game in games:
         x.append(game[2:12])
     return x
-
 
 def get_dependent(games: list) -> list:
     y = []
@@ -34,13 +29,11 @@ def get_dependent(games: list) -> list:
 
     return y
 
-
 def get_info(games: list) -> list:
     info = []
     for game in games:
         info.append(game[0:3])
     return info
-
 
 def split_data():
     training_games = get_games("Data/PregameStats.csv")
@@ -60,24 +53,9 @@ def split_data():
     return x_training, y_training, x_curr_season, y_curr_season
 
 
-def train_forrest_classifier(x, y):
-    # Train Model: Fitting Random Forest Classification to the Training set
-    rf = RandomForestClassifier(n_estimators=100, criterion='entropy', random_state=42)
-    rf.fit(x, y)
-
-    return rf
-
-
-def train_neural_net(x, y):
-    clf = MLPClassifier(solver='lbfgs', alpha=1e-5, hidden_layer_sizes=(5, 2), random_state=1)
-    clf.fit(x, y)
-    return clf
-
-
 def predict_results(classifier, x, y):
     y_prediction = classifier.predict(x)
     print(accuracy_score(y, y_prediction))
-
 
 def clean_x(x):
     cleaned_x = []
@@ -90,27 +68,25 @@ def clean_x(x):
 
     return cleaned_x
 
+
 # Best is 55% with forrest and no data standardization
 def train_model():
     x_training_headers = ['win_ratio_5', 'draw_ratio_5', 'loss_ratio_5', 'h2h_w_d_l_ratio', 'current_goals_avg',
                   'current_goals_avg_h_a', 'conceded_goals_avg', 'conceded_goals_avg_h_a', 'goal_average_5',
                   'conceded_average_5']
     x_train, y_train, x_curr_season, y_curr_season = split_data()
-    # x_train = clean_x(x_train)
-    # x_curr_season = clean_x(x_curr_season)
 
-    # TODO: Split methods into separate file (predictor creators)
     # Look into cleaning data and what is best for each predictor
     # Combine predictions (sensor fusion)
 
     # Make predictions on original Dimensionality
-    forrest_classifier = train_forrest_classifier(x_train, y_train)
+    forrest_classifier = models.train_forrest_classifier(x_train, y_train)
     predict_results(forrest_classifier, x_curr_season, y_curr_season)
     print("----------------------------------------------")
-    neural_net_classifier = train_neural_net(x_train, y_train)
+    neural_net_classifier = models.train_neural_net(x_train, y_train)
     predict_results(neural_net_classifier, x_curr_season, y_curr_season)
     print("----------------------------------------------")
-    kmeans = KMeans(n_clusters=2, random_state=0).fit(x_train, y_train)
+    kmeans = models.train_kmeans(x_train, y_train)
     predict_results(kmeans, x_curr_season, y_curr_season)
     print("----------------------------------------------")
 
@@ -120,7 +96,7 @@ def train_model():
     x_curr_season_reduced = pca.fit_transform(x_curr_season)
 
     # Predict Reduced Dimensionality
-    kmeans = KMeans(n_clusters=2, random_state=0).fit(x_train_reduced, y_train)
+    kmeans = models.train_kmeans(x_train_reduced, y_train)
     predict_results(kmeans, x_curr_season_reduced, y_curr_season)
 
 train_model()
